+++ jodd-json/src/main/java/jodd/json/impl/BooleanArrayJsonSerializer.java
@@ -35,6 +35,10 @@ public class BooleanArrayJsonSerializer implements TypeJsonSerializer<boolean[]>
 
 	@Override
 	public boolean serialize(final JsonContext jsonContext, final boolean[] array) {
+		if (array.length == 0 && jsonContext.isExcludeEmpty()) {
+			return true;
+		}
+
 		jsonContext.writeOpenArray();
 
 		for (int i = 0; i < array.length; i++) {