+++ jodd-json/src/main/java/jodd/json/impl/IntArrayJsonSerializer.java
@@ -35,6 +35,10 @@ public class IntArrayJsonSerializer implements TypeJsonSerializer<int[]> {
 
 	@Override
 	public boolean serialize(final JsonContext jsonContext, final int[] array) {
+		if (array.length == 0 && jsonContext.isExcludeEmpty()) {
+			return true;
+		}
+		
 		jsonContext.writeOpenArray();
 
 		for (int i = 0; i < array.length; i++) {