+++ jodd-json/src/main/java/jodd/json/impl/MapJsonSerializer.java
@@ -37,6 +37,9 @@ public class MapJsonSerializer extends KeyValueJsonSerializer<Map<?, ?>> {
 
 	@Override
 	public void serializeValue(final JsonContext jsonContext, final Map<?, ?> map) {
+		if (map.isEmpty() && jsonContext.isExcludeEmpty()) {
+			return;
+		}
 		jsonContext.writeOpenObject();
 
 		int count = 0;